Override `repr_html_batch of `BBoxFeature`
