+++ tensorflow_datasets/core/features/bounding_boxes.py
@@ -17,8 +17,11 @@
 
 import collections
 
+import numpy as np
 import tensorflow.compat.v2 as tf
 
+from tensorflow_datasets.core import utils
+from tensorflow_datasets.core import lazy_imports_lib
 from tensorflow_datasets.core.features import feature
 from tensorflow_datasets.core.utils import type_utils
 
@@ -78,6 +81,31 @@ class BBoxFeature(feature.Tensor):
     return super(BBoxFeature, self).encode_example(
         [bbox.ymin, bbox.xmin, bbox.ymax, bbox.xmax]
     )
+
+  def _build_thumbnail_with_bbox(self, ex: np.ndarray) -> str:
+    """Returns the HTML str representation of an Image with BBoxes."""
+    PIL_Image = lazy_imports_lib.lazy_imports.PIL_Image
+    PIL_ImageDraw = lazy_imports_lib.lazy_imports.PIL_ImageDraw
+
+    SIZE = 150
+    blank_img = PIL_Image.new('RGB', (SIZE, SIZE), (255, 255, 255))
+    draw = PIL_ImageDraw.Draw(blank_img)
+
+    for i in range(ex.shape[0]):
+      # Rescale coordinates to match size of blank_image
+      ymin, xmin, ymax, xmax = ex[i, :]*SIZE
+      # Generate random rgb values for Bbox ouline
+      r, g, b = list(np.random.randint(0, 256, size=3))
+      draw.rectangle(((xmin, ymin), (xmax, ymax)), outline=(r, g, b))
+    return blank_img
+
+  def repr_html_batch(self, ex: np.ndarray) -> str:
+    """Returns the HTML str representation of the object."""
+    img = self._build_thumbnail_with_bbox(ex)
+
+    img_str = utils.get_base64(lambda buff: img.save(buff, format='PNG'))
+
+    return f'<img src="data:image/png;base64,{img_str}" alt="Img" />'
 
   @classmethod
   def from_json_content(cls, value: Json) -> 'BBoxFeature':
